% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eclthemes.R
\name{theme_ecl}
\alias{theme_ecl}
\title{A clean theme based on ggthemes::theme_fivethirtyeight, but with a white
background, axis titles, and a vertical legend on the right.}
\usage{
theme_ecl(base_size = 16, base_family = "sans")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
A clean theme based on ggthemes::theme_fivethirtyeight, but with a white
background, axis titles, and a vertical legend on the right.
}
\examples{
(ggplot(mtcars, aes(mpg, color=factor(cyl), fill=factor(cyl), group=factor(cyl))))
 + geom_density(alpha=0.5)
 + theme_ecl()
}

